+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/combobox/DefaultItemComboBoxBuilder.java
@@ -173,10 +173,10 @@ final class DefaultItemComboBoxBuilder<T> extends AbstractComponentBuilder<T, JC
 			comboBox.setEditor(editor);
 		}
 		if (mouseWheelScrolling) {
-			comboBox.addMouseWheelListener(new ComboBoxMouseWheelListener(itemComboBoxModel, false));
+			comboBox.addMouseWheelListener(new ComboBoxMouseWheelListener(comboBox, false));
 		}
 		if (mouseWheelScrollingWithWrapAround) {
-			comboBox.addMouseWheelListener(new ComboBoxMouseWheelListener(comboBoxModel, true));
+			comboBox.addMouseWheelListener(new ComboBoxMouseWheelListener(comboBox, true));
 		}
 		if (maximumRowCount >= 0) {
 			comboBox.setMaximumRowCount(maximumRowCount);