+++ changelog.md
@@ -52,6 +52,7 @@ Codion Change Log
 - ProgressWorker.Builder.onCancelled() handler now gets called in case the background task throws CancelException.
 - DefaultProgressWorkerDialogBuilder, CancelException handling removed, some refactoring.
 - FilterTable bug fixed, sorting actions no longer enabled for columns with sorting locked in the sort model.
+- ComboBoxMouseWheelListener and SpinnerMouseWheelListener bug fixed, no longer enabled when component is disabled.
 ### is.codion.framework.domain
 - EntityDefinition.description() now returns Optional.
 - DefaultEntityDefinition, source attributes validated for denormalized attributes.